Use the underlying message to throw MppingExcp
Instead of using "unknown error" can we be using the exception message of the previous exception so that the original message can be retrieved from the exception's getMessage method. The use case being, say we have a case class defined with some assertions on the fields, while converting the json to an object, it fails throwing IllegalArgumentException, now if we need to pass on the actual message about the field for which the validation has failed, that wont be possible anymore as the "unknown error" message gets fetched if we call the getMessage method.